requirements.txt
edge_tts==7.0.0
fastapi==0.115.7
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.41.1
tqdm==4.67.1
uvicorn==0.34.0
